refactor: tasking module structure, naming, and UI
