+++ community/src/main/java/org/dhis2/community/tasking/repositories/TaskingModule.kt
@@ -1,19 +1,21 @@
-package org.dhis2.community.tasks
+package org.dhis2.community.tasking.repositories
 
 import dagger.Module
 import dagger.Provides
 import javax.inject.Singleton
+import org.dhis2.community.tasking.ui.TaskingView
+import org.dhis2.community.tasking.engine.TaskingPresenter
 
 @Module
-class TasksModule(private val view: TaskView) {
+class TaskingModule(private val view: TaskingView) {
 
     @Provides
     @Singleton
-    fun provideView(): TaskView = view
+    fun provideView(): TaskingView = view
 
     @Provides
     @Singleton
-    fun providePresenter(view: TaskView): TasksPresenter {
-        return TasksPresenter(view)
+    fun providePresenter(view: TaskingView): TaskingPresenter {
+        return TaskingPresenter(view)
     }
 }
