--- community/src/main/java/org/dhis2/community/tasking/ui/TaskView.kt
@@ -1,6 +1,0 @@
-package org.dhis2.community.tasking.ui
-
-interface TaskView {
-    fun showSyncDialog()
-}
-
