--- community/src/main/java/org/dhis2/community/tasking/ui/TasksPresenter.kt
@@ -1,7 +1,0 @@
-package org.dhis2.community.tasking.ui
-
-class TasksPresenter(private val view: TaskView) {
-    fun onResume() {
-    }
-}
-
